fix: updates from review
