+++ packages/ap-core/alliance_platform/core/util.py
@@ -1,0 +1,16 @@
+# Taken verbatim from distutils source in Python 3.11
+# https://github.com/python/cpython/blob/v3.11.2/Lib/distutils/util.py#L308
+def strtobool(val):
+    """Convert a string representation of truth to true (1) or false (0).
+
+    True values are 'y', 'yes', 't', 'true', 'on', and '1'; false values
+    are 'n', 'no', 'f', 'false', 'off', and '0'.  Raises ValueError if
+    'val' is anything else.
+    """
+    val = val.lower()
+    if val in ("y", "yes", "t", "true", "on", "1"):
+        return 1
+    elif val in ("n", "no", "f", "false", "off", "0"):
+        return 0
+    else:
+        raise ValueError("invalid truth value %r" % (val,))
