+++ packages/ap-pdf/docs/installation.rst
@@ -18,17 +18,6 @@ Add ``alliance_platform.pdf``, ``alliance_platform.frontend``, and ``alliance_pl
         ...
     ]
 
-Heroku
-~~~~~~
-
-* Add the https://github.com/Thomas-Boi/heroku-playwright-python-browsers.git buildpack and set `PLAYWRIGHT_BUILDPACK_BROWSERS=chromium` in the Heroku settings.
-
-* Add the https://github.com/playwright-community/heroku-playwright-buildpack buildpack
-
-* ``CHROMIUM_EXECUTABLE_PATH`` can be set to the chromium path if needed. This will be set by the Heroku buildpack automatically. In local dev this isn't necessary.
-
-Note that the chromium executable and dependencies will increase the base slug size.
-
 
 Settings
 ~~~~~~~~
